Fix updating of RT/Fave state on RTs
If `A` is reblogged in `B` and we reblog `B` to make `C` then the
returned value has `C` referencing `A`, with the new "is reblogged"
status. By making `C` and `B` listen to the status on `A` then we can
update both Posts when `A` is updated by reblogging either one.